Adjusted epoxy kit:     Edited product names to match marketing product names
