+++ Calculator/ListTest/Select Line/Select Line Epoxy/BCoatSelectTS.swift
@@ -8,14 +8,18 @@
 import SwiftUI
 import Combine
 
+// ONLY COMMERCIAL RESIN BEIGE, CLEAR, LIGHT GRAY, AND PT B TH
 
+var selectTSABC_TS = stats(product: "Top Shelf® Epoxy Part A:", covRate: 420, MixRat: 2)
+var selectTSBBC_TS = stats(product: "Top Shelf® Epoxy Part B:", covRate: 420, MixRat: 1)
+// var SelectTSColorantBC_TS = stats(product: "Top Shelf® Epoxy Colorant:", covRate: 210, MixRat: 16)
 
-struct PCoatSelectTS: View {
+struct BCoatSelectTS: View {
     
  
     @EnvironmentObject var ChipTS : ColorChipValuesSelectTS //Used for coat selections
-    @EnvironmentObject var selectTSA_TS : stats //Used for system stats (mixRat,CovRate, etc)
-    @EnvironmentObject var selectTSB_TS : stats //Used for system stats (mixRat,CovRate, etc)
+    @EnvironmentObject var TSA_TS : stats //Used for system stats (mixRat,CovRate, etc)
+    @EnvironmentObject var TSB_TS : stats //Used for system stats (mixRat,CovRate, etc)
     @EnvironmentObject var TSColorant_TS : stats //Used for system stats (mixRat,CovRate, etc)
     @EnvironmentObject var SolventCleaner : stats 
     @EnvironmentObject var sf : SquareFeet //Square feet
@@ -24,7 +28,7 @@ struct PCoatSelectTS: View {
     let kitTypes = ["Clear", "Light Gray", "Beige"]
     
     let kitCodes = ["EX-102001-KIT", "EX-10401-KIT", "EX-10101-KIT"]
-    
+
     let TSColorantChoices = ["No Color (pre-pigmented Part A)", "Beige", "Black", "Dark Gray", "Enchanted Green", "Handicap Blue", "Latte", "Light Gray", "Medium Gray", "Mocha", "Safety Blue", "Safety Red", "Safety Yellow", "Shadow Gray", "Tan", "Tile Red", "White",]
     
     let TSColorantCodes = ["No Color (pre-pigmented Part A)", "EX-KTSECLBG-EA", "EX-KTSECLBL-EA", "EX-KTSECLDG-EA", "EX-KTSECLEG-EA", "EX-KTSECLHB-EA", "EX-KTSECLLT-EA", "EX-KTSECLLG-EA", "EX-KTSECLMG-EA", "EX-KTSECLMC-EA", "EX-KTSECLSB-EA", "EX-KTSECLSR-EA", "EX-KTSECLSY-EA", "EX-KTSECLSG-EA", "EX-KTSECLTN-EA", "EX-KTSECLTR-EA", "EX-KTSECLWH-01",]
@@ -35,22 +39,21 @@ struct PCoatSelectTS: View {
          return kit
         } // returns number of kits required
     
-  
-    
     var body: some View {
         VStack {
             HStack {
-                Text("Coating Thickness: 3-5 mils")
+                Text("Coating Thickness: 8-12 mils")
                     .fontWeight(.heavy)
                 Spacer()
-                ChipTSSelectPCoatInfo()
+                ChipTSSelectBCoatInfo()
             }
             .padding()
+            
            Text("Choose your color:")
             .fontWeight(.bold)
-            Picker(selection: $ChipTS.PCoatPtA,
+            Picker(selection: $ChipTS.BCoatPtA,
                 label: ZStack {
-                    Text("\(kitTypes[ChipTS.PCoatPtA])")
+                    Text("\(kitTypes[ChipTS.BCoatPtA])")
                         .opacity(1)
                     Rectangle()
                         .fill(Color.gray)
@@ -65,12 +68,15 @@ struct PCoatSelectTS: View {
             }
             .background(Color.gray.opacity(0.2))
             .cornerRadius(5)
-
+            
+//             .onChange(of: ChipTS.PCoatPtB) { _ in
+//                quant(product: TSB)
+//             }
 //            Text("Top Shelf Epoxy® Colorant:")
 //             .fontWeight(.bold)
-//             Picker(selection: $ChipTS.PCoatTSColorant,
+//             Picker(selection: $ChipTS.BCoatTSColorant,
 //                 label: ZStack {
-//                     Text("\(TSColorantChoices[ChipTS.PCoatTSColorant])")
+//                     Text("\(TSColorantChoices[ChipTS.BCoatTSColorant])")
 //                         .opacity(1)
 //                     Rectangle()
 //                         .fill(Color.gray)
@@ -101,41 +107,53 @@ struct PCoatSelectTS: View {
                 }
                 
                 HStack {
-                        Text("\(kitCodes[ChipTS.PCoatPtA])")
+                        Text("\(kitCodes[ChipTS.BCoatPtA])")
                             .font(.caption)
                         Spacer()
-                    Text("Select Line Kit - \(kitTypes[ChipTS.PCoatPtA])")
+                    Text("Kretus Select Epoxy Kit - \(kitTypes[ChipTS.BCoatPtA])")
                             .font(.caption)
                         Spacer()
-                        Text("\(quant(product: selectTSA_TS))")
+                        Text("\(quant(product: selectTSABC_TS))")
                             .font(.caption)
                     }
             
-//                if ChipTS.PCoatTSColorant != 0 {
+//                HStack { // Solvent Cleaner
+//                    Text("EX-KSLVCLN-EA")
+//                        .font(.caption)
+//                    Spacer()
+//                    Text("Solvent Cleaner, 32 oz")
+//                        .font(.caption)
+//                    Spacer()
+//                    Text("\(quant(product: selectTSABC_TS))")
+//                        .font(.caption)
+//                }
+//                if ChipTS.BCoatTSColorant != 0 {
 //                    HStack {
-//                        Text("\(TSColorantCodes[ChipTS.PCoatTSColorant])")
+//                        Text("\(TSColorantCodes[ChipTS.BCoatTSColorant])")
 //                            .font(.caption)
 //                        Spacer()
-//                        Text("Top Shelf® Epoxy Colorant: \(TSColorantChoices[ChipTS.PCoatTSColorant]), 16 oz")
+//                        Text("Top Shelf® Epoxy Colorant: \(TSColorantChoices[ChipTS.BCoatTSColorant]), 16 oz")
 //                            .font(.caption)
 //                        Spacer()
-//                        Text("\(quant(product: TSColorant_TS))")
+//                        Text("\(quant(product: TSColorantBC_TS))")
 //                            .font(.caption)
-//                            }
+//                    }
 //                
-//                    if ChipTS.PCoatTSColorant == 4 { //USED TO ABBREVIATE ENCHANTED GREEN TO FIT ON SCREEN
+//                    if ChipTS.BCoatTSColorant == 4 { //USED TO ABBREVIATE ENCHANTED GREEN TO FIT ON SCREEN
 //                        HStack {
-//                            Text("\(TSColorantCodes[ChipTS.PCoatTSColorant])")
+//                            Text("\(TSColorantCodes[ChipTS.BCoatTSColorant])")
 //                                .font(.caption)
 //                            Spacer()
 //                            Text("TSE Colorant:  Ench. Green, 16 oz")
 //                                .font(.caption)
 //                            Spacer()
-//                            Text("\(quant(product: TSColorant_TS))")
+//                            Text("\(quant(product: TSColorantBC_TS))")
 //                                .font(.caption)
-//                                }
-//                                                    }
-//                                                }
+//                        }
+//                    }
+//
+//                    
+//            }
         }
             .background(Color(red:239.0/255.0, green: 243.0/255.0, blue: 244.0/250, opacity: 1.0))
             .cornerRadius(5.0)
@@ -143,7 +161,7 @@ struct PCoatSelectTS: View {
             HStack {
                 Text("Add Waste Factor: ")
                 Spacer()
-                TextField("",value: $ChipTS.PCoatWaste, formatter: NumberFormatter())
+                TextField("",value: $ChipTS.BCoatWaste, formatter: NumberFormatter())
                     .frame(width:30, height:25)
                     .background(Color(red:239.0/255.0, green: 243.0/255.0, blue: 244.0/250, opacity: 1.0))
                     .cornerRadius(5.0)
@@ -153,14 +171,14 @@ struct PCoatSelectTS: View {
             HStack {
                 Text("Total:")
                 Spacer()
-                Text("\(quant(product: selectTSA_TS) + ChipTS.PCoatWaste) kit(s)")
+                Text("\(quant(product: selectTSABC_TS) + ChipTS.BCoatWaste) kit(s)")
             }
             .padding()
         }
     }
 }
 
-struct PCoatSelectTS_Previews: PreviewProvider {
+struct BCoatSelectTS_Previews: PreviewProvider {
     static var previews: some View {
         BCoatGarage()
     }
